Work on the ASCII I/O implementation
ASCII I/O is nearing completion. Current implementation is alpha
quality. The writer part is certainly not going to be optimal in terms
of performance, we would need `fmt`'s number to string formatting for
starters.

The current interface is a bit cumbersome, and may need additional
refactoring. It now assumes that the elements in the PLY file can be
directly mapped on to some data type property for property, or at least
casted property for property. So in case a client type contains
additional data interleaved with the PLY element data, it is not
possible to directly read data from the PLY file into the user's type.
This can be solved by adding stride support though, there are some ideas
on how to implement that.

Also, more exhaustive tests need to be added still.