Switch from aggregate to non-aggregate types
Convert `PlyProperty` and `PlyElement` from semi-aggregate types to
non-aggregate types; that is, add various constructors and member
property accessors. In C++11, a struct for which some of the properties
have a default initializer in the struct definition can not be aggregate
types.

This simplifies adding properties to elements a little bit. Furthermore,
the boolean `isList` on a `PlyProperty` becomes an implementation detail
which is no longer exposed through its interface.