Fix bug in element seek logic
Fix an issue where seeking to the start of element data would fail in
case the input stream read EOF before. It is required to call
`std::istream::clear()` prior to `std::istream::seekg()` to fix this.

Adds `PlyElement::copyWithSizeHint` as a kind of ugly convenience
function to get an element definition with a size hint for a certain
property.