Restructure project a bit
All private include headers are now stored in a separate sub-directory
'plywoot', as is done in Boost projects for example. Furthermore, all
symbols that do not need to be visible outside the public interface are
now in the 'detail' namespace, again, similar to how Boost handles this.

Furthermore, update `.clang-format` to deal with nested namespaces in a
nicer way.