Implement simple reflection system
Implement a simple reflection type system and use it on the write side
for now.

This allows to precisely specify how objects of some type in the source
data map onto properties in the PLY file.

This requires to always declare how the source data structure is laid
out in memory. The automatic deduction code path is removed, and this
will be done for the read side as well. This will require the user to
always declare how the data in memory maps onto data on disk, which may
be less convenient, but always results in the fastest code path being
taken. Furthermore, it reduces the size of the API.

I am not ruling out that automatic deduction of the memory layout will
be re-implemented in the future, but right now I do not have a need for
it.