Implement support for binary PLY files
Only binary little endian mode is supported for now. The main limitation
right now for binary little endian mode read support is that the order
in which elements are read in should match the order in which they are
defined in the input file. In other words, out-of-order element
retrieval is not supported.

The reason for this is that it is not trivial to skip over elements in
binary mode in an efficient manner in case the element contains a list
property of variable size without having specified some size hint either
through a type map for the list (we only support fixed size arrays...),
or through some size hint that we got rid of earlier...