Change to a stateful API
Remove the requirement on the input streams that are passed to
`plywoot::IStream` to be seekable. This restricts the API in the sense
that elements need to be iterated and evaluated in the order as they
appear in the input stream; out-of-order retrieval of element data is no
longer possible.

This makes the API a bit more ugly in my opinion, with the big advantage
that non-seekable streams can now be passed to `plywoot::IStream` as
well. For example, this allows PLYwoot to be used with Boost's filtering
streams to read PLY data from compressed data streams.

Finally, this results in a minor performance improvement as well, since
we now guarantee that input data is only buffered once.