Fix TODO (actually a bug)
When writing elements, the number of input elements specified as input
to the write function determines the number of elements as stored in the
PLY header, instead of the size property that happens to be associated
with the PLY element passed to the write function.

This prevents the user from having to specify a size that matches the
number of input items when creating the `PlyElement` instance that
describes the data to write.