Fix incorrect memcpy handling of list properties
List properties mapping on arrays in memory may never be memcpy'd, since
the data in the PLY file contains a size per list entry, which needs to
be skipped. The list itself is still memcpy'd, per element, instead of
all elements being memcpy'd in one go.

Furthermore, fix incorrect handling of the number of properties mapping
to an array in memory; this should always be one property, a list
property. Somehow mapping multiple properties on to an array should
never be done; using a `reflect::Pack` for that.