Implement `plywoot::convert`
Implement `plywoot::convert` which converts a PLY input stream of a
certain format to another PLY format, and outputs all data to a given
output stream.

This requires element read functionality that does not depend on a
specific element layout specified at compile time. This is implemented
using an overload for `plywoot::readElement()` returning an instance of
`PlyElementData`, which wraps a block of memory representing an
anonymous struct of the PLY data in the parsed PLY data satisfying all
alignment requirements of the target platform. For now, `PlyElementData`
is only used by `plywoot::convert` and as such lacks certain functions
to access the underlying data in a convenient manner. This may be added
in the future.