Improve endianness handling
Do not depend on `<endian.h>` for byte order swapping functionality,
improves compatibility with e.g. Windows.

Furthermore, do not make any assumptions on the endianness of the target
platform, instead, determine the endianness of the platform at compile
time, and only swap around bytes in case the byte order of the PLY file
does not match the endianness of the platform for which PLYwoot was
compiled.