Simplify `reflect::Layout`
Only require construction of a `reflect::Layout` instance when writing
data to a PLY file. When reading PLY data, the `reflect::Layout` type is
merely used as a vehicle to propagate the list of types in the layout.

This removes the need to store a non-const data pointer in the layout
type.